+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 - Fixed first paragraph check returning both positive and negative
 - Snippet description now has minimum height
 
+### Improved
+- It's now possible to pass social media meta to `craft.seo.custom`
+
 ## 3.0.1 - 2018-01-23
 ### Fixed
 - Fixed error when installing on Craft Personal